TCOTA-160 Reorganized so that hasStorageAccess check is lighter weight
This change was also done to replace the HTTP redirect. I thought that the
session cookie wasn't getting set in Safari Private Browsing because of the HTTP
redirect, but even with this approach a session cookie isn't getting set
in Safari Private Browsing.

Still, since this change makes the initial check lighter weight, I'm going to keep it.